Refactor invocation of old pass manager into legacy function
